Fix Reader handling for OffsetOutOfRange (#891)
Provide a new configuration `AutoOffsetReset` that can be used to
reset the non-group-id reader's offset when the topic has been
recreated (or the current offset is no longer valid for any reason).
The `AutoOffsetReset` can have a value of 0 (no action), or
FirstOffset / LastOffset.

Added a new test case in `TestReader` to verify that, on setting the
offset initially past the last offset in the topic, the reader will
restart from the beginning once it detects this OffsetOutOfRange
error, instead of perpetually reporting OffsetOutOfRange and waiting
for the producer to catch up.

Fixes #891.