Add "human readable" description for offset in reader logs
When the reader logs an offset value, if the value is a special value,
i.e. FirstOffset / LastOffset, it will log something like

    setting the offset ... from -2 to 20

With this change, it'll say "first offset" or "last offset",
respectively, for example:

    setting the offset ... from first offset to 20

It uses a new `fmt.Formatter` type, `humanOffset`, which seems more
expressive (and performant) that simply converting to a string.

Instead of type-casting offset directly, I added a helper
`toHumanOffset` to avoid other types, e.g. `humanOffset(uint64(1))`.